+++ src/modules/tools/indexTools.js
@@ -10,11 +10,12 @@ import actions from "./actionsTools";
 import Draw from "./draw/store/indexDraw";
 import FileImport from "./fileImport/store/indexFileImport";
 import Gfi from "./gfi/store/indexGfi";
-import SaveSelection from "./saveSelection/store/indexSaveSelection";
-import ScaleSwitcher from "./scaleSwitcher/store/indexScaleSwitcher";
-import SupplyCoord from "./supplyCoord/store/indexSupplyCoord";
-import SearchByCoord from "./searchByCoord/store/indexSearchByCoord";
 import Measure from "./measure/store/indexMeasure";
+import SaveSelection from "./saveSelection/store/indexSaveSelection";
+import SearchByCoord from "./searchByCoord/store/indexSearchByCoord";
+import ScaleSwitcher from "./scaleSwitcher/store/indexScaleSwitcher";
+import StyleVT from "./styleVT/store/indexStyleVT";
+import SupplyCoord from "./supplyCoord/store/indexSupplyCoord";
 
 /**
  * This is here to test app-store/utils/composeModules.
@@ -26,11 +27,12 @@ export default {
         Draw,
         FileImport,
         Gfi,
+        Measure,
         SaveSelection,
-        ScaleSwitcher,
-        SupplyCoord,
         SearchByCoord,
-        Measure
+        ScaleSwitcher,
+        StyleVT,
+        SupplyCoord
     },
     state,
     getters,